+++ src/docs/bootstrap.ts
@@ -1,6 +1,6 @@
 require('es6-shim');
-require('zone.js');
 require('reflect-metadata');
+require('zone.js');
 require('highlight.js/styles/atelier-estuary-light.css');
 
 import {LocationStrategy, HashLocationStrategy} from '@angular/common';
@@ -11,6 +11,7 @@ import {bootstrap} from '@angular/platform-browser-dynamic';
 import {App} from './app.component';
 import {APP_ROUTER_PROVIDERS} from './app.routes';
 import {logTemplateErrors} from './development-tools';
+import {DragStateTrackerFactory} from '../components/file-drop-area/drag-state-tracker.service';
 
 bootstrap(App, [
     APP_ROUTER_PROVIDERS,
@@ -18,6 +19,7 @@ bootstrap(App, [
     provideForms(),
     logTemplateErrors(),
     Title,
+    DragStateTrackerFactory,
     { provide: LocationStrategy, useClass: HashLocationStrategy }
 ])
 .catch(err => console.error(err));