+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixes
 * SplitViewContainer works with OnPush change detection
 * SplitViewContainer handles dragging outside of the window
+* Select uses ellipses if selected item text is wider than component
 
 ## 0.8.0 (2016-08-24)